Improve the Framework configuration and its bootstrap stage
