+++ system/bootstrap.php
@@ -10,16 +10,25 @@
 use Core\Aliases;
 use Core\Router;
 use Helpers\Session;
-use App\Config;
 
 
 /** Turn on output buffering. */
 ob_start();
 
-/** Initialize the Configuration */
-Config::init();
+/** Load the Configuration. */
+require APPDIR .'Config.php';
 
-/** Initialize the Aliases */
+/** Set the Default Timezone. */
+date_default_timezone_set(DEFAULT_TIMEZONE);
+
+/** Turn on the custom error handling. */
+set_exception_handler('Core\Logger::ExceptionHandler');
+set_error_handler('Core\Logger::ErrorHandler');
+
+/** Load the Framework wide functions. */
+require dirname(__FILE__) .DS .'functions.php';
+
+/** Initialize the Aliases. */
 Aliases::init();
 
 /** Start the Session. */
@@ -27,9 +36,6 @@ Session::init();
 
 /** Initialize the Language. */
 Language::init();
-
-/** Load the Framework wide functions. */
-require dirname(__FILE__) .DS .'functions.php';
 
 /** Get the Router instance. */
 $router = Router::getInstance();
